Begin splitting out most database operations to BrainzStem
